+++ backend/__mocks__/fileMock.ts
@@ -1,0 +1,1 @@
+module.exports = 'test-file-stub';
