+++ backend/src/index.ts
@@ -1,12 +1,10 @@
 import 'reflect-metadata';
 import app from './express';
-import dataSource from './dataSource';
 import { logMessages } from './utils';
-
-export * from './modules';
+import { dataSource } from './database/dataSource';
 
 const { API_PORT, DEBUG } = process.env;
-console.log('port:', API_PORT);
+
 const {
   SERVER_START,
   UTC_DATE_TIME,
@@ -16,29 +14,20 @@ const {
   DATABASE_CONNECTION_ERROR,
 } = logMessages;
 
-// Connect to the database and initialize it.
 dataSource
   .initialize()
-  .then(async () => {
+  .then(() => {
     console.info(DATABASE_CONNECTION);
-
-    // Start the Express application (server).
-    try {
-      app.listen(Number(API_PORT), () => {
-        // Log server start information.
-        console.info(SERVER_START);
-        console.info(CURRENT_NODE_VERSION);
-        console.info(UTC_DATE_TIME);
-        console.info(PACIFIC_DATE_TIME);
-        if (DEBUG)
-          console.info("DEBUG is enabled. Set environment variable to 'false' to disable.");
-      });
-    } catch (error) {
-      // Log any error that occurs during the server start.
-      console.error(error);
-    }
+    app.listen(Number(API_PORT), () => {
+      // Log server start information.
+      console.info(SERVER_START);
+      console.info(CURRENT_NODE_VERSION);
+      console.info(UTC_DATE_TIME);
+      console.info(PACIFIC_DATE_TIME);
+      if (DEBUG) console.info("DEBUG is enabled. Set environment variable to 'false' to disable.");
+    });
   })
   .catch((error) => {
-    // Log an error if the database connection and initialization fails.
-    console.error(DATABASE_CONNECTION_ERROR, error);
+    console.error(DATABASE_CONNECTION_ERROR);
+    console.error(error);
   });
