+++ backend/src/modules/user/router.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import dataSource from '../../dataSource';
+import { dataSource } from '../../database/dataSource';
 import { createController, createRepository, createService } from '../common';
 import { User } from './entity';
