+++ backend/tsconfig.json
@@ -17,7 +17,8 @@
     }
   },
   "include": [
-    "**/*"
+      "src",
+      "__mocks__",
   ],
   "exclude": [
     "tsconfig.json",
