+++ frontend/src/hooks/useAuthentication/useAuthentication.ts
@@ -1,19 +1,9 @@
-import {
-  KeycloakProvider,
-  useKeycloak,
-  decodeJWT,
-} from '@bcgov/citz-imb-kc-react';
+import { KeycloakProvider, useKeycloak, decodeJWT } from '@bcgov/citz-imb-kc-react';
 import { useMemo } from 'react';
 
 export const useAuthentication = () => {
-  const {
-    getAuthorizationHeaderValue,
-    hasRole,
-    login,
-    logout,
-    refreshToken,
-    state,
-  } = useKeycloak();
+  const { getAuthorizationHeaderValue, hasRole, login, logout, refreshToken, state } =
+    useKeycloak();
 
   const isAuthenticated = useMemo(() => {
     return state?.userInfo !== undefined;
